Add functionality to fetch only new entries
inspiretools.blg2texkey parses a BibTeX log to check which entries
aren't in the database yet and output the texkeys, much as
inspiretools.aux2texkey does.

bin/blgtobib uses this and inspiretools.texkey2bib to download only
BibTeX entries which aren't already found in a database. It should be
given a .blg file.

Example usage:

    blgtobib example.blg >> bibliography.bib

This should make it easier to add just a few citations to a large
document.  As a bonus, new entries appear at the end of the .bib, making
version control of the database cleaner.

Todo: biber support.